@@ -14,58 +14,77 @@
   "files": [
     "src",
     "dist/*.js",
+    "dist/*.css",
+    "dist/locales/*.json",
     "types/*.d.ts"
   ],
   "scripts": {
     "dev": "rollup -w -c build/config.js --environment TARGET:web-umd-dev",
     "dev:cjs": "rollup -w -c build/config.js --environment TARGET:web-cjs",
     "dev:esm": "rollup -w -c build/config.js --environment TARGET:web-esm",
-    "build": "node build/build.js",
+    "bundle": "node build/build.js",
+    "build": "npm run bundle && webpack -p",
     "build:umd": "rollup -w -c build/config.js --environment TARGET:web-umd-dev",
-    "lint": "eslint src/**/*.js tests/**/*.js --no-fix",
-    "test": "jest"
+    "build:amd": "webpack -p",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "test": "npm run e2e && npm run unit",
+    "unit": "jest tests/unit/",
+    "e2e": "node tests/e2e/samples.js test",
+    "e2e:update": "node tests/e2e/samples.js update",
+    "build:samples": "node samples/source/index.js generate"
   },
   "dependencies": {
-    "promise-polyfill": "8.1.0",
-    "svg.draggable.js": "^2.2.1",
+    "svg.draggable.js": "^2.2.2",
     "svg.easing.js": "^2.0.0",
     "svg.filter.js": "^2.0.2",
-    "svg.js": "^2.6.6",
     "svg.pathmorphing.js": "^0.1.3",
-    "svg.resize.js": "^1.4.1",
+    "svg.resize.js": "^1.4.3",
     "svg.select.js": "^2.1.2"
   },
   "devDependencies": {
-    "@babel/core": "7.3.4",
-    "@babel/plugin-proposal-class-properties": "7.2.3",
-    "@babel/preset-env": "7.3.4",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "10.0.1",
-    "babel-jest": "24.1.0",
-    "chalk": "2.4.2",
-    "cross-env": "5.2.0",
-    "eslint": "5.14.1",
-    "eslint-config-prettier": "3.3.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-prettier": "3.0.1",
-    "eslint-plugin-promise": "4.0.1",
-    "jest": "24.1.0",
-    "jest-puppeteer": "3.8.0",
-    "prettier": "1.16.4",
-    "puppeteer": "1.11.0",
-    "puppeteer-core": "1.11.0",
-    "rollup": "1.1.0",
-    "rollup-plugin-babel": "4.3.2",
-    "rollup-plugin-copy-glob": "0.3.0",
-    "rollup-plugin-json": "3.1.0",
-    "rollup-plugin-node-resolve": "4.0.0",
+    "@babel/core": "7.7.7",
+    "@babel/plugin-proposal-class-properties": "7.7.4",
+    "@babel/preset-env": "7.7.7",
+    "babel-eslint": "10.0.3",
+    "babel-jest": "24.9.0",
+    "babel-loader": "8.0.6",
+    "babel-plugin-istanbul": "6.0.0",
+    "chalk": "3.0.0",
+    "css-loader": "3.4.1",
+    "eslint": "6.8.0",
+    "eslint-config-prettier": "6.9.0",
+    "eslint-loader": "3.0.3",
+    "eslint-plugin-import": "2.19.1",
+    "eslint-plugin-prettier": "3.1.2",
+    "eslint-plugin-promise": "4.2.1",
+    "fs-extra": "8.1.0",
+    "jest": "24.9.0",
+    "nunjucks": "3.2.0",
+    "nyc": "15.0.0",
+    "pixelmatch": "5.1.0",
+    "pngjs": "3.4.0",
+    "prettier": "1.19.1",
+    "puppeteer": "2.0.0",
+    "puppeteer-cluster": "0.18.0",
+    "@rollup/plugin-json": "4.0.1",
+    "@rollup/plugin-node-resolve": "6.0.0",
+    "@rollup/plugin-replace": "2.3.0",
+    "@rollup/plugin-strip": "1.3.1",
+    "rollup": "1.27.14",
+    "rollup-plugin-babel": "4.3.3",
+    "rollup-plugin-copy-glob": "0.3.1",
     "rollup-plugin-postcss": "2.0.3",
-    "rollup-plugin-replace": "2.1.0",
-    "rollup-plugin-strip": "1.2.1",
-    "rollup-plugin-svgo": "1.0.2",
-    "terser": "3.14.1",
+    "rollup-plugin-svgo": "1.1.0",
+    "rollup-plugin-terser": "5.1.3",
+    "style-loader": "1.1.2",
+    "svg-inline-loader": "0.8.0",
+    "terser": "4.4.3",
     "tslint": "5.12.1",
-    "typescript": "3.2.2"
+    "typescript": "3.2.2",
+    "webpack": "4.41.5",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-cli": "^3.3.10"
   },
   "author": {
     "name": "Juned Chhipa",
