@@ -17,6 +17,8 @@ export default class Animations {
   setEasingFunctions() {
     let easing
 
+    if (this.w.globals.easing) return
+
     const userDefinedEasing = this.w.config.chart.animations.easing
 
     switch (userDefinedEasing) {
@@ -38,22 +40,25 @@ export default class Animations {
       }
       case 'swing': {
         easing = (pos) => {
-          var s = 1.70158
-          return (pos -= 1) * pos * ((s + 1) * pos + s) + 1
+          let s = 1.70158
+          let ret = (pos -= 1) * pos * ((s + 1) * pos + s) + 1
+          return ret
         }
         break
       }
       case 'bounce': {
         easing = (pos) => {
+          let ret = ''
           if (pos < 1 / 2.75) {
-            return 7.5625 * pos * pos
+            ret = 7.5625 * pos * pos
           } else if (pos < 2 / 2.75) {
-            return 7.5625 * (pos -= 1.5 / 2.75) * pos + 0.75
+            ret = 7.5625 * (pos -= 1.5 / 2.75) * pos + 0.75
           } else if (pos < 2.5 / 2.75) {
-            return 7.5625 * (pos -= 2.25 / 2.75) * pos + 0.9375
+            ret = 7.5625 * (pos -= 2.25 / 2.75) * pos + 0.9375
           } else {
-            return 7.5625 * (pos -= 2.625 / 2.75) * pos + 0.984375
+            ret = 7.5625 * (pos -= 2.625 / 2.75) * pos + 0.984375
           }
+          return ret
         }
         break
       }
@@ -86,7 +91,7 @@ export default class Animations {
   /*
    ** Animate radius of a circle element
    */
-  animateCircleRadius(el, from, to, speed, easing) {
+  animateCircleRadius(el, from, to, speed, easing, cb) {
     if (!from) from = 0
 
     el.attr({
@@ -95,6 +100,9 @@ export default class Animations {
       .animate(speed, easing)
       .attr({
         r: to
+      })
+      .afterAll(() => {
+        cb()
       })
   }
 
@@ -122,13 +130,11 @@ export default class Animations {
     el.attr(from)
       .animate(speed)
       .attr(to)
-      .afterAll(function() {
-        fn()
-      })
+      .afterAll(() => fn())
   }
 
   animatePathsGradually(params) {
-    let { el, j, pathFrom, pathTo, speed, delay, strokeWidth } = params
+    let { el, realIndex, j, fill, pathFrom, pathTo, speed, delay } = params
 
     let me = this
     let w = this.w
@@ -143,16 +149,19 @@ export default class Animations {
       w.config.chart.animations.dynamicAnimation.enabled &&
       w.globals.dataChanged
     ) {
-      delayFactor = 0
+      // disabled due to this bug - https://github.com/apexcharts/vue-apexcharts/issues/75
+      // delayFactor = 0
     }
-
     me.morphSVG(
       el,
+      realIndex,
       j,
+      w.config.chart.type === 'line' && !w.globals.comboCharts
+        ? 'stroke'
+        : fill,
       pathFrom,
       pathTo,
       speed,
-      strokeWidth,
       delay * delayFactor
     )
   }
@@ -160,12 +169,23 @@ export default class Animations {
   showDelayedElements() {
     this.w.globals.delayedElements.forEach((d) => {
       const ele = d.el
-      ele.classList.remove('hidden')
+      ele.classList.remove('apexcharts-element-hidden')
     })
   }
 
+  animationCompleted(el) {
+    const w = this.w
+    if (w.globals.animationEnded) return
+
+    w.globals.animationEnded = true
+
+    if (typeof w.config.chart.events.animationEnd === 'function') {
+      w.config.chart.events.animationEnd(this.ctx, { el, w })
+    }
+  }
+
   // SVG.js animation for morphing one path to another
-  morphSVG(el, j, pathFrom, pathTo, speed, strokeWidth, delay) {
+  morphSVG(el, realIndex, j, fill, pathFrom, pathTo, speed, delay) {
     let w = this.w
 
     if (!pathFrom) {
@@ -182,11 +202,9 @@ export default class Animations {
       pathFrom.indexOf('NaN') > -1
     ) {
       pathFrom = `M 0 ${w.globals.gridHeight}`
-      speed = 1
     }
     if (pathTo.indexOf('undefined') > -1 || pathTo.indexOf('NaN') > -1) {
       pathTo = `M 0 ${w.globals.gridHeight}`
-      speed = 1
     }
     if (!w.globals.shouldAnimate) {
       speed = 1
@@ -205,12 +223,15 @@ export default class Animations {
             j === w.globals.series[w.globals.maxValsInArrayIndex].length - 2 &&
             w.globals.shouldAnimate
           ) {
-            w.globals.animationEnded = true
+            this.animationCompleted(el)
           }
-        } else if (w.globals.shouldAnimate) {
-          w.globals.animationEnded = true
-          if (typeof w.config.chart.events.animationEnd === 'function') {
-            w.config.chart.events.animationEnd(this.ctx, w)
+        } else if (fill !== 'none' && w.globals.shouldAnimate) {
+          if (
+            (!w.globals.comboCharts &&
+              realIndex === w.globals.series.length - 1) ||
+            w.globals.comboCharts
+          ) {
+            this.animationCompleted(el)
           }
         }
 
