@@ -3,19 +3,18 @@ import BarStacked from '../charts/BarStacked'
 import CandleStick from '../charts/CandleStick'
 import CoreUtils from './CoreUtils'
 import Crosshairs from './Crosshairs'
-import DateTime from './../utils/DateTime'
 import HeatMap from '../charts/HeatMap'
+import Globals from '../modules/settings/Globals'
 import Pie from '../charts/Pie'
 import Radar from '../charts/Radar'
 import Radial from '../charts/Radial'
+import RangeBar from '../charts/RangeBar'
+import Legend from './legend/Legend'
 import Line from '../charts/Line'
 import Graphics from './Graphics'
-import XAxis from './axes/XAxis'
-import YAxis from './axes/YAxis'
 import Range from './Range'
 import Utils from '../utils/Utils'
 import Scales from './Scales'
-import Series from './Series'
 import TimeScale from './TimeScale'
 
 /**
@@ -29,12 +28,6 @@ export default class Core {
     this.ctx = ctx
     this.w = ctx.w
     this.el = el
-
-    this.coreUtils = new CoreUtils(this.ctx)
-
-    this.twoDSeries = []
-    this.threeDSeries = []
-    this.twoDSeriesX = []
   }
 
   // get data and store into appropriate vars
@@ -50,10 +43,12 @@ export default class Core {
       'line',
       'area',
       'bar',
+      'rangeBar',
+      // 'rangeArea',
       'candlestick',
-      'radar',
       'scatter',
       'bubble',
+      'radar',
       'heatmap'
     ]
 
@@ -61,6 +56,8 @@ export default class Core {
       'line',
       'area',
       'bar',
+      'rangeBar',
+      // 'rangeArea',
       'candlestick',
       'scatter',
       'bubble'
@@ -69,6 +66,10 @@ export default class Core {
     gl.axisCharts = axisChartsArrTypes.indexOf(ct) > -1
 
     gl.xyCharts = xyChartsArrTypes.indexOf(ct) > -1
+
+    gl.isBarHorizontal =
+      (cnf.chart.type === 'bar' || cnf.chart.type === 'rangeBar') &&
+      cnf.plotOptions.bar.horizontal
 
     gl.chartClass = '.apexcharts' + gl.cuid
 
@@ -144,9 +145,15 @@ export default class Core {
       // if user has specified a particular type for particular series
       if (typeof ser[st].type !== 'undefined') {
         if (ser[st].type === 'column' || ser[st].type === 'bar') {
-          w.config.plotOptions.bar.horizontal = false // bar not supported in mixed charts
+          if (gl.series.length > 1 && cnf.plotOptions.bar.horizontal) {
+            // horizontal bars not supported in mixed charts, hence show a warning
+            console.warn(
+              'Horizontal bars are not supported in a mixed/combo chart. Please turn off `plotOptions.bar.horizontal`'
+            )
+          }
           columnSeries.series.push(series)
           columnSeries.i.push(st)
+          w.globals.columnSeries = columnSeries.series
         } else if (ser[st].type === 'area') {
           areaSeries.series.push(series)
           areaSeries.i.push(st)
@@ -179,6 +186,7 @@ export default class Core {
     let candlestick = new CandleStick(this.ctx, xyRatios)
     let pie = new Pie(this.ctx)
     let radialBar = new Radial(this.ctx)
+    let rangeBar = new RangeBar(this.ctx, xyRatios)
     let radar = new Radar(this.ctx)
     let elGraph = []
 
@@ -209,13 +217,12 @@ export default class Core {
           scatterLine.draw(scatterSeries.series, 'scatter', scatterSeries.i)
         )
       }
-      // TODO: allow bubble series in a combo chart
-      // if (bubbleSeries.series.length > 0) {
-      //   const bubbleLine = new Line(this.ctx, xyRatios, true)
-      //   elGraph.push(
-      //     bubbleLine.draw(bubbleSeries.series, 'bubble', bubbleSeries.i)
-      //   )
-      // }
+      if (bubbleSeries.series.length > 0) {
+        const bubbleLine = new Line(this.ctx, xyRatios, true)
+        elGraph.push(
+          bubbleLine.draw(bubbleSeries.series, 'bubble', bubbleSeries.i)
+        )
+      }
     } else {
       switch (cnf.chart.type) {
         case 'line':
@@ -236,6 +243,9 @@ export default class Core {
         case 'candlestick':
           let candleStick = new CandleStick(this.ctx, xyRatios)
           elGraph = candleStick.draw(gl.series)
+          break
+        case 'rangeBar':
+          elGraph = rangeBar.draw(gl.series)
           break
         case 'heatmap':
           let heatmap = new HeatMap(this.ctx, xyRatios)
@@ -279,10 +289,10 @@ export default class Core {
           elDim = Utils.getDimensions(this.el.parentNode)
         }
 
-        gl.svgWidth = (elDim[0] * parseInt(cnf.chart.width)) / 100
+        gl.svgWidth = (elDim[0] * parseInt(cnf.chart.width, 10)) / 100
       }
     } else if (widthUnit === 'px' || widthUnit === '') {
-      gl.svgWidth = parseInt(cnf.chart.width)
+      gl.svgWidth = parseInt(cnf.chart.width, 10)
     }
 
     if (gl.svgHeight !== 'auto' && gl.svgHeight !== '') {
@@ -292,17 +302,20 @@ export default class Core {
         .pop()
       if (heightUnit === '%') {
         let elParentDim = Utils.getDimensions(this.el.parentNode)
-        gl.svgHeight = (elParentDim[1] * parseInt(cnf.chart.height)) / 100
+        gl.svgHeight = (elParentDim[1] * parseInt(cnf.chart.height, 10)) / 100
       } else {
-        gl.svgHeight = parseInt(cnf.chart.height)
+        gl.svgHeight = parseInt(cnf.chart.height, 10)
       }
     } else {
       if (gl.axisCharts) {
         gl.svgHeight = gl.svgWidth / 1.61
       } else {
-        gl.svgHeight = gl.svgWidth
+        gl.svgHeight = gl.svgWidth / 1.2
       }
     }
+
+    if (gl.svgWidth < 0) gl.svgWidth = 0
+    if (gl.svgHeight < 0) gl.svgHeight = 0
 
     Graphics.setAttrs(gl.dom.Paper.node, {
       width: gl.svgWidth,
@@ -310,7 +323,11 @@ export default class Core {
     })
 
     // gl.dom.Paper.node.parentNode.parentNode.style.minWidth = gl.svgWidth + "px";
-    let offsetY = cnf.chart.sparkline.enabled ? 0 : gl.axisCharts ? 14 : 5
+    let offsetY = cnf.chart.sparkline.enabled
+      ? 0
+      : gl.axisCharts
+      ? cnf.chart.parentHeightOffset
+      : 0
 
     gl.dom.Paper.node.parentNode.parentNode.style.minHeight =
       gl.svgHeight + offsetY + 'px'
@@ -329,6 +346,56 @@ export default class Core {
       transform: 'translate(' + tX + ', ' + tY + ')'
     }
     Graphics.setAttrs(gl.dom.elGraphical.node, scalingAttrs)
+
+    gl.x2SpaceAvailable = gl.svgWidth - gl.dom.elGraphical.x() - gl.gridWidth
+  }
+
+  // To prevent extra spacings in the bottom of the chart, we need to recalculate the height for pie/donut/radialbar charts
+  resizeNonAxisCharts() {
+    const w = this.w
+
+    const gl = w.globals
+
+    let legendHeight = 0
+    let offY = w.config.chart.sparkline.enabled ? 0 : 15
+    offY = offY + w.config.grid.padding.bottom
+
+    if (
+      (w.config.legend.position === 'top' ||
+        w.config.legend.position === 'bottom') &&
+      w.config.legend.show &&
+      !w.config.legend.floating
+    ) {
+      legendHeight =
+        new Legend(this.ctx).legendHelpers.getLegendBBox().clwh + 10
+    }
+
+    let radialEl = w.globals.dom.baseEl.querySelector('.apexcharts-radialbar')
+
+    let chartInnerDimensions = w.globals.radialSize * 2.05
+
+    if (radialEl && !w.config.chart.sparkline.enabled) {
+      let elRadialRect = Utils.getBoundingClientRect(radialEl)
+      chartInnerDimensions = elRadialRect.bottom
+
+      let maxHeight = elRadialRect.bottom - elRadialRect.top
+
+      chartInnerDimensions = Math.max(w.globals.radialSize * 2.05, maxHeight)
+    }
+
+    let newHeight = chartInnerDimensions + gl.translateY + legendHeight + offY
+
+    if (gl.dom.elLegendForeign) {
+      gl.dom.elLegendForeign.setAttribute('height', newHeight)
+    }
+
+    gl.dom.elWrap.style.height = newHeight + 'px'
+
+    Graphics.setAttrs(gl.dom.Paper.node, {
+      height: newHeight
+    })
+
+    gl.dom.Paper.node.parentNode.parentNode.style.minHeight = newHeight + 'px'
   }
 
   /*
@@ -340,60 +407,15 @@ export default class Core {
   }
 
   resetGlobals() {
+    const resetxyValues = () => {
+      return this.w.config.series.map((s) => [])
+    }
+    const globalObj = new Globals()
+
     let gl = this.w.globals
-    gl.series = []
-    gl.seriesCandleO = []
-    gl.seriesCandleH = []
-    gl.seriesCandleL = []
-    gl.seriesCandleC = []
-    gl.seriesPercent = []
-    gl.seriesX = []
-    gl.seriesZ = []
-    gl.seriesNames = []
-    gl.seriesTotals = []
-    gl.stackedSeriesTotals = []
-    gl.labels = []
-    gl.timelineLabels = []
-    gl.noLabelsProvided = false
-    gl.timescaleTicks = []
-    gl.resizeTimer = null
-    gl.selectionResizeTimer = null
-    gl.seriesXvalues = (() => {
-      return this.w.config.series.map((s) => {
-        return []
-      })
-    })()
-    gl.seriesYvalues = (() => {
-      return this.w.config.series.map((s) => {
-        return []
-      })
-    })()
-    gl.delayedElements = []
-    gl.pointsArray = []
-    gl.dataLabelsRects = []
-    gl.isXNumeric = false
-    gl.isDataXYZ = false
-    gl.maxY = -Number.MAX_VALUE
-    gl.minY = Number.MIN_VALUE
-    gl.minYArr = []
-    gl.maxYArr = []
-    gl.maxX = -Number.MAX_VALUE
-    gl.minX = Number.MAX_VALUE
-    gl.initialmaxX = -Number.MAX_VALUE
-    gl.initialminX = Number.MAX_VALUE
-    gl.maxDate = 0
-    gl.minDate = Number.MAX_VALUE
-    gl.minZ = Number.MAX_VALUE
-    gl.maxZ = -Number.MAX_VALUE
-    gl.yAxisScale = []
-    gl.xAxisScale = null
-    gl.xAxisTicksPositions = []
-    gl.yLabelsCoords = []
-    gl.yTitleCoords = []
-    gl.xRange = 0
-    gl.yRange = []
-    gl.zRange = 0
-    gl.dataPoints = 0
+    globalObj.initGlobalVars(gl)
+    gl.seriesXvalues = resetxyValues()
+    gl.seriesYvalues = resetxyValues()
   }
 
   isMultipleY() {
@@ -402,406 +424,8 @@ export default class Core {
       this.w.config.yaxis.constructor === Array &&
       this.w.config.yaxis.length > 1
     ) {
-      // first, turn off stacking if multiple y axis
-      this.w.config.chart.stacked = false
       this.w.globals.isMultipleYAxis = true
       return true
-    }
-  }
-
-  excludeCollapsedSeriesInYAxis() {
-    const w = this.w
-    w.globals.ignoreYAxisIndexes = w.globals.collapsedSeries.map(
-      (collapsed, i) => {
-        if (this.w.globals.isMultipleYAxis) {
-          return collapsed.index
-        }
-      }
-    )
-  }
-
-  isMultiFormat() {
-    return this.isFormatXY() || this.isFormat2DArray()
-  }
-
-  // given format is [{x, y}, {x, y}]
-  isFormatXY() {
-    const series = this.w.config.series.slice()
-
-    const sr = new Series(this.ctx)
-    this.activeSeriesIndex = sr.getActiveConfigSeriesIndex()
-
-    if (
-      typeof series[this.activeSeriesIndex].data !== 'undefined' &&
-      series[this.activeSeriesIndex].data.length > 0 &&
-      series[this.activeSeriesIndex].data[0] !== null &&
-      typeof series[this.activeSeriesIndex].data[0].x !== 'undefined' &&
-      series[this.activeSeriesIndex].data[0] !== null
-    ) {
-      return true
-    }
-  }
-
-  // given format is [[x, y], [x, y]]
-  isFormat2DArray() {
-    const series = this.w.config.series.slice()
-
-    const sr = new Series(this.ctx)
-    this.activeSeriesIndex = sr.getActiveConfigSeriesIndex()
-
-    if (
-      typeof series[this.activeSeriesIndex].data !== 'undefined' &&
-      series[this.activeSeriesIndex].data.length > 0 &&
-      typeof series[this.activeSeriesIndex].data[0] !== 'undefined' &&
-      series[this.activeSeriesIndex].data[0] !== null &&
-      series[this.activeSeriesIndex].data[0].constructor === Array
-    ) {
-      return true
-    }
-  }
-
-  handleFormat2DArray(ser, i) {
-    const cnf = this.w.config
-    const gl = this.w.globals
-
-    for (let j = 0; j < ser[i].data.length; j++) {
-      if (typeof ser[i].data[j][1] !== 'undefined') {
-        if (
-          Array.isArray(ser[i].data[j][1]) &&
-          ser[i].data[j][1].length === 4
-        ) {
-          this.twoDSeries.push(Utils.parseNumber(ser[i].data[j][1][3]))
-        } else {
-          this.twoDSeries.push(Utils.parseNumber(ser[i].data[j][1]))
-        }
-      }
-      if (cnf.xaxis.type === 'datetime') {
-        // if timestamps are provided and xaxis type is datettime,
-
-        let ts = new Date(ser[i].data[j][0])
-        ts = new Date(ts).getTime()
-        this.twoDSeriesX.push(ts)
-      } else {
-        this.twoDSeriesX.push(ser[i].data[j][0])
-      }
-    }
-
-    for (let j = 0; j < ser[i].data.length; j++) {
-      if (typeof ser[i].data[j][2] !== 'undefined') {
-        this.threeDSeries.push(ser[i].data[j][2])
-        gl.isDataXYZ = true
-      }
-    }
-  }
-
-  handleFormatXY(ser, i) {
-    const cnf = this.w.config
-    const gl = this.w.globals
-
-    const dt = new DateTime(this.ctx)
-
-    let activeI = i
-    if (gl.collapsedSeriesIndices.indexOf(i) > -1) {
-      // fix #368
-      activeI = this.activeSeriesIndex
-    }
-
-    // get series
-    for (let j = 0; j < ser[i].data.length; j++) {
-      if (typeof ser[i].data[j].y !== 'undefined') {
-        if (Array.isArray(ser[i].data[j].y) && ser[i].data[j].y.length === 4) {
-          this.twoDSeries.push(Utils.parseNumber(ser[i].data[j].y[3]))
-        } else {
-          this.twoDSeries.push(Utils.parseNumber(ser[i].data[j].y))
-        }
-      }
-    }
-
-    // get seriesX
-    for (let j = 0; j < ser[activeI].data.length; j++) {
-      const isXString = typeof ser[activeI].data[j].x === 'string'
-      const isXDate = !!dt.isValidDate(ser[activeI].data[j].x.toString())
-
-      if (isXString || isXDate) {
-        // user supplied '01/01/2017' or a date string (a JS date object is not supported)
-        if (isXString) {
-          if (cnf.xaxis.type === 'datetime') {
-            this.twoDSeriesX.push(dt.parseDate(ser[activeI].data[j].x))
-          } else {
-            // a category and not a numeric x value
-            this.fallbackToCategory = true
-            this.twoDSeriesX.push(ser[activeI].data[j].x)
-          }
-        } else {
-          if (cnf.xaxis.type === 'datetime') {
-            this.twoDSeriesX.push(
-              dt.parseDate(ser[activeI].data[j].x.toString())
-            )
-          } else {
-            this.twoDSeriesX.push(parseFloat(ser[activeI].data[j].x))
-          }
-        }
-      } else {
-        // a numeric value in x property
-        this.twoDSeriesX.push(ser[activeI].data[j].x)
-      }
-    }
-
-    if (ser[i].data[0] && typeof ser[i].data[0].z !== 'undefined') {
-      for (let t = 0; t < ser[i].data.length; t++) {
-        this.threeDSeries.push(ser[i].data[t].z)
-      }
-      gl.isDataXYZ = true
-    }
-  }
-
-  handleCandleStickData(ser, i) {
-    const gl = this.w.globals
-
-    let ohlc = {}
-    if (this.isFormat2DArray()) {
-      ohlc = this.handleCandleStickDataFormat('array', ser, i)
-    } else if (this.isFormatXY()) {
-      ohlc = this.handleCandleStickDataFormat('xy', ser, i)
-    }
-
-    gl.seriesCandleO.push(ohlc.o)
-    gl.seriesCandleH.push(ohlc.h)
-    gl.seriesCandleL.push(ohlc.l)
-    gl.seriesCandleC.push(ohlc.c)
-
-    return ohlc
-  }
-
-  handleCandleStickDataFormat(format, ser, i) {
-    const serO = []
-    const serH = []
-    const serL = []
-    const serC = []
-
-    const err =
-      'Please provide [Open, High, Low and Close] values in valid format. Read more https://apexcharts.com/docs/series/#candlestick'
-
-    if (format === 'array') {
-      if (ser[i].data[0][1].length !== 4) {
-        throw new Error(err)
-      }
-      for (let j = 0; j < ser[i].data.length; j++) {
-        serO.push(ser[i].data[j][1][0])
-        serH.push(ser[i].data[j][1][1])
-        serL.push(ser[i].data[j][1][2])
-        serC.push(ser[i].data[j][1][3])
-      }
-    } else if (format === 'xy') {
-      if (ser[i].data[0].y.length !== 4) {
-        throw new Error(err)
-      }
-      for (let j = 0; j < ser[i].data.length; j++) {
-        serO.push(ser[i].data[j].y[0])
-        serH.push(ser[i].data[j].y[1])
-        serL.push(ser[i].data[j].y[2])
-        serC.push(ser[i].data[j].y[3])
-      }
-    }
-
-    return {
-      o: serO,
-      h: serH,
-      l: serL,
-      c: serC
-    }
-  }
-
-  parseDataAxisCharts(ser, ctx = this.ctx) {
-    const cnf = this.w.config
-    const gl = this.w.globals
-
-    const dt = new DateTime(ctx)
-
-    for (let i = 0; i < ser.length; i++) {
-      this.twoDSeries = []
-      this.twoDSeriesX = []
-      this.threeDSeries = []
-
-      if (typeof ser[i].data === 'undefined') {
-        console.error(
-          "It is a possibility that you may have not included 'data' property in series."
-        )
-        return
-      }
-
-      if (this.isMultiFormat()) {
-        if (this.isFormat2DArray()) {
-          this.handleFormat2DArray(ser, i)
-        } else if (this.isFormatXY()) {
-          this.handleFormatXY(ser, i)
-        }
-
-        if (cnf.chart.type === 'candlestick' || ser[i].type === 'candlestick') {
-          this.handleCandleStickData(ser, i)
-        }
-
-        gl.series.push(this.twoDSeries)
-        gl.labels.push(this.twoDSeriesX)
-        gl.seriesX.push(this.twoDSeriesX)
-
-        if (!this.fallbackToCategory) {
-          gl.isXNumeric = true
-        }
-      } else {
-        if (cnf.xaxis.type === 'datetime') {
-          // user didn't supplied [{x,y}] or [[x,y]], but single array in data.
-          // Also labels/categories were supplied differently
-          gl.isXNumeric = true
-          const dates =
-            cnf.labels.length > 0
-              ? cnf.labels.slice()
-              : cnf.xaxis.categories.slice()
-
-          for (let j = 0; j < dates.length; j++) {
-            if (typeof dates[j] === 'string') {
-              let isDate = dt.isValidDate(dates[j])
-              if (isDate) {
-                this.twoDSeriesX.push(dt.parseDate(dates[j]))
-              } else {
-                throw new Error(
-                  'You have provided invalid Date format. Please provide a valid JavaScript Date'
-                )
-              }
-            }
-          }
-
-          gl.seriesX.push(this.twoDSeriesX)
-        } else if (cnf.xaxis.type === 'numeric') {
-          gl.isXNumeric = true
-          let x =
-            cnf.labels.length > 0
-              ? cnf.labels.slice()
-              : cnf.xaxis.categories.slice()
-
-          if (x.length > 0) {
-            this.twoDSeriesX = x
-            gl.seriesX.push(this.twoDSeriesX)
-          }
-        }
-        gl.labels.push(this.twoDSeriesX)
-        const singleArray = ser[i].data.map((d) => {
-          return Utils.parseNumber(d)
-        })
-        gl.series.push(singleArray)
-      }
-
-      gl.seriesZ.push(this.threeDSeries)
-
-      if (ser[i].name !== undefined) {
-        gl.seriesNames.push(ser[i].name)
-      } else {
-        gl.seriesNames.push('series-' + parseInt(i + 1))
-      }
-    }
-
-    return this.w
-  }
-
-  parseDataNonAxisCharts(ser) {
-    const gl = this.w.globals
-    const cnf = this.w.config
-
-    gl.series = ser.slice()
-    gl.seriesNames = cnf.labels.slice()
-    for (let i = 0; i < gl.series.length; i++) {
-      if (gl.seriesNames[i] === undefined) {
-        gl.seriesNames.push('series-' + (i + 1))
-      }
-    }
-
-    return this.w
-  }
-
-  handleExternalLabelsData(ser) {
-    const cnf = this.w.config
-    const gl = this.w.globals
-
-    // user provided labels in category axis
-    if (cnf.xaxis.categories.length > 0) {
-      gl.labels = cnf.xaxis.categories
-    } else if (cnf.labels.length > 0) {
-      gl.labels = cnf.labels.slice()
-    } else if (this.fallbackToCategory) {
-      gl.labels = gl.labels[0]
-    } else {
-      // user didn't provided labels, fallback to 1-2-3-4-5
-      let labelArr = []
-      if (gl.axisCharts) {
-        for (let i = 0; i < gl.series[gl.maxValsInArrayIndex].length; i++) {
-          labelArr.push(i + 1)
-        }
-
-        for (let i = 0; i < ser.length; i++) {
-          gl.seriesX.push(labelArr)
-        }
-
-        gl.isXNumeric = true
-      }
-
-      // no series to pull labels from, put a 0-10 series
-      if (labelArr.length === 0) {
-        labelArr = [0, 10]
-        for (let i = 0; i < ser.length; i++) {
-          gl.seriesX.push(labelArr)
-        }
-      }
-
-      gl.labels = labelArr
-      gl.noLabelsProvided = true
-
-      if (cnf.xaxis.type === 'category') {
-        gl.isXNumeric = false
-      }
-    }
-  }
-
-  // Segregate user provided data into appropriate vars
-  parseData(ser) {
-    let w = this.w
-    let cnf = w.config
-    let gl = w.globals
-    this.excludeCollapsedSeriesInYAxis()
-
-    this.fallbackToCategory = false
-
-    this.resetGlobals()
-    this.isMultipleY()
-
-    if (gl.axisCharts) {
-      this.parseDataAxisCharts(ser)
-    } else {
-      this.parseDataNonAxisCharts(ser)
-    }
-
-    this.coreUtils.getLargestSeries()
-
-    // set Null values to 0 in all series when user hides/shows some series
-    if (cnf.chart.type === 'bar' && cnf.chart.stacked) {
-      const series = new Series(this.ctx)
-      gl.series = series.setNullSeriesToZeroValues(gl.series)
-    }
-
-    this.coreUtils.getSeriesTotals()
-    if (gl.axisCharts) {
-      this.coreUtils.getStackedSeriesTotals()
-    }
-
-    this.coreUtils.getPercentSeries()
-
-    // user didn't provide a [[x,y],[x,y]] series, but a named series
-    if (
-      !gl.isXNumeric ||
-      (cnf.xaxis.type === 'numeric' &&
-        cnf.labels.length === 0 &&
-        cnf.xaxis.categories.length === 0)
-    ) {
-      this.handleExternalLabelsData(ser)
     }
   }
 
@@ -819,59 +443,34 @@ export default class Core {
         crosshairs.drawYCrosshairs()
       }
 
-      xyRatios = this.coreUtils.getCalculatedRatios()
-
       if (
         w.config.xaxis.type === 'datetime' &&
-        w.config.xaxis.labels.formatter === undefined &&
-        isFinite(w.globals.minX) &&
-        isFinite(w.globals.maxX)
+        w.config.xaxis.labels.formatter === undefined
       ) {
         let ts = new TimeScale(this.ctx)
-        const formattedTimeScale = ts.calculateTimeScaleTicks(
-          w.globals.minX,
-          w.globals.maxX
-        )
+        let formattedTimeScale = []
+        if (
+          isFinite(w.globals.minX) &&
+          isFinite(w.globals.maxX) &&
+          !w.globals.isBarHorizontal
+        ) {
+          formattedTimeScale = ts.calculateTimeScaleTicks(
+            w.globals.minX,
+            w.globals.maxX
+          )
+        } else if (w.globals.isBarHorizontal) {
+          formattedTimeScale = ts.calculateTimeScaleTicks(
+            w.globals.minY,
+            w.globals.maxY
+          )
+        }
         ts.recalcDimensionsBasedOnFormat(formattedTimeScale)
       }
+
+      const coreUtils = new CoreUtils(this.ctx)
+      xyRatios = coreUtils.getCalculatedRatios()
     }
     return xyRatios
-  }
-
-  drawAxis(type, xyRatios) {
-    let gl = this.w.globals
-    let cnf = this.w.config
-
-    let xAxis = new XAxis(this.ctx)
-    let yAxis = new YAxis(this.ctx)
-
-    if (gl.axisCharts && type !== 'radar') {
-      let elXaxis, elYaxis
-
-      if (type === 'bar' && cnf.plotOptions.bar.horizontal) {
-        elYaxis = yAxis.drawYaxisInversed(0)
-        elXaxis = xAxis.drawXaxisInversed(0)
-
-        gl.dom.elGraphical.add(elXaxis)
-        gl.dom.elGraphical.add(elYaxis)
-      } else {
-        elXaxis = xAxis.drawXaxis()
-        gl.dom.elGraphical.add(elXaxis)
-
-        cnf.yaxis.map((yaxe, index) => {
-          if (gl.ignoreYAxisIndexes.indexOf(index) === -1) {
-            elYaxis = yAxis.drawYaxis(index)
-            gl.dom.Paper.add(elYaxis)
-          }
-        })
-      }
-    }
-
-    cnf.yaxis.map((yaxe, index) => {
-      if (gl.ignoreYAxisIndexes.indexOf(index) === -1) {
-        yAxis.yAxisTitleRotate(index, yaxe.opposite)
-      }
-    })
   }
 
   setupBrushHandler() {
@@ -894,7 +493,7 @@ export default class Core {
         targetChart.w.globals.brushSource = this.ctx
 
         let updateSourceChart = () => {
-          this.ctx._updateOptions(
+          this.ctx.updateHelpers._updateOptions(
             {
               chart: {
                 selection: {
@@ -919,27 +518,39 @@ export default class Core {
             updateSourceChart()
           }
         }
+      })
 
-        w.config.chart.events.selection = (chart, e) => {
+      w.config.chart.events.selection = (chart, e) => {
+        targets.forEach((target) => {
+          let targetChart = ApexCharts.getChartByID(target)
           let yaxis = Utils.clone(w.config.yaxis)
-          if (w.config.chart.brush.autoScaleYaxis) {
+
+          if (
+            w.config.chart.brush.autoScaleYaxis &&
+            targetChart.w.globals.series.length === 1
+          ) {
             const scale = new Scales(targetChart)
-            yaxis = scale.autoScaleY(targetChart, e)
+            yaxis = scale.autoScaleY(targetChart, yaxis, e)
           }
-          targetChart._updateOptions(
+          targetChart.ctx.updateHelpers._updateOptions(
             {
               xaxis: {
                 min: e.xaxis.min,
                 max: e.xaxis.max
               },
-              yaxis
+              yaxis: {
+                ...targetChart.w.config.yaxis[0],
+                min: yaxis[0].min,
+                max: yaxis[0].max
+              }
             },
+            false,
             false,
             false,
             false
           )
-        }
-      })
+        })
+      }
     }
   }
 }
