@@ -10,20 +10,25 @@ class CoreUtils {
 
   static checkComboSeries(series) {
     let comboCharts = false
-    let comboChartsHasBars = false
+    let comboBarCount = 0
+
     // if user specified a type in series too, turn on comboCharts flag
     if (series.length && typeof series[0].type !== 'undefined') {
       comboCharts = true
       series.forEach((s) => {
-        if (s.type === 'bar' || s.type === 'column') {
-          comboChartsHasBars = true
+        if (
+          s.type === 'bar' ||
+          s.type === 'column' ||
+          s.type === 'candlestick'
+        ) {
+          comboBarCount++
         }
       })
     }
 
     return {
-      comboCharts,
-      comboChartsHasBars
+      comboBarCount,
+      comboCharts
     }
   }
 
@@ -37,6 +42,8 @@ class CoreUtils {
     const w = this.w
     let total = []
 
+    if (w.globals.series.length === 0) return total
+
     for (
       let i = 0;
       i < w.globals.series[w.globals.maxValsInArrayIndex].length;
@@ -44,7 +51,9 @@ class CoreUtils {
     ) {
       let t = 0
       for (let j = 0; j < w.globals.series.length; j++) {
-        t += w.globals.series[j][i]
+        if (typeof w.globals.series[j][i] !== 'undefined') {
+          t += w.globals.series[j][i]
+        }
       }
       total.push(t)
     }
@@ -56,14 +65,10 @@ class CoreUtils {
   getSeriesTotalByIndex(index = null) {
     if (index === null) {
       // non-plot chart types - pie / donut / circle
-      return this.w.config.series.reduce((acc, cur) => {
-        return acc + cur
-      }, 0)
+      return this.w.config.series.reduce((acc, cur) => acc + cur, 0)
     } else {
       // axis charts - supporting multiple series
-      return this.w.globals.series[index].reduce((acc, cur) => {
-        return acc + cur
-      }, 0)
+      return this.w.globals.series[index].reduce((acc, cur) => acc + cur, 0)
     }
   }
 
@@ -71,38 +76,39 @@ class CoreUtils {
     let r = []
     if (index === null) {
       // non-plot chart types - pie / donut / circle
-      r = this.w.config.series.filter((d) => {
-        return d !== null
-      })
+      r = this.w.config.series.filter((d) => d !== null)
     } else {
       // axis charts - supporting multiple series
-      r = this.w.globals.series[index].filter((d) => {
-        return d !== null
-      })
+      r = this.w.globals.series[index].filter((d) => d !== null)
     }
 
     return r.length === 0
   }
 
   seriesHaveSameValues(index) {
-    return this.w.globals.series[index].every((val, i, arr) => {
-      return val === arr[0]
-    })
+    return this.w.globals.series[index].every((val, i, arr) => val === arr[0])
   }
 
+  getCategoryLabels(labels) {
+    const w = this.w
+    let catLabels = labels.slice()
+    if (w.config.xaxis.convertedCatToNumeric) {
+      catLabels = labels.map((i) => {
+        return w.config.xaxis.labels.formatter(i - w.globals.minX + 1)
+      })
+    }
+
+    return catLabels
+  }
   // maxValsInArrayIndex is the index of series[] which has the largest number of items
   getLargestSeries() {
     const w = this.w
     w.globals.maxValsInArrayIndex = w.globals.series
-      .map(function(a) {
-        return a.length
-      })
+      .map((a) => a.length)
       .indexOf(
         Math.max.apply(
           Math,
-          w.globals.series.map(function(a) {
-            return a.length
-          })
+          w.globals.series.map((a) => a.length)
         )
       )
   }
@@ -111,7 +117,7 @@ class CoreUtils {
     const w = this.w
     let size = 0
 
-    w.globals.markers.size.forEach(function(m) {
+    w.globals.markers.size.forEach((m) => {
       size = Math.max(size, m)
     })
 
@@ -179,14 +185,15 @@ class CoreUtils {
       let seriesPercent = []
       if (Array.isArray(ser)) {
         for (let j = 0; j < ser.length; j++) {
-          const total = w.globals.stackedSeriesTotals[j]
-          let percent = (100 * ser[j]) / total
+          let total = w.globals.stackedSeriesTotals[j]
+          let percent = 0
+          if (total) {
+            percent = (100 * ser[j]) / total
+          }
           seriesPercent.push(percent)
         }
       } else {
-        const total = w.globals.seriesTotals.reduce((acc, val) => {
-          return acc + val
-        }, 0)
+        const total = w.globals.seriesTotals.reduce((acc, val) => acc + val, 0)
         let percent = (100 * ser) / total
         seriesPercent.push(percent)
       }
@@ -226,11 +233,16 @@ class CoreUtils {
     }
 
     xRatio = gl.xRange / gl.gridWidth
-    initialXRatio = Math.abs(gl.initialmaxX - gl.initialminX) / gl.gridWidth
+
+    initialXRatio = Math.abs(gl.initialMaxX - gl.initialMinX) / gl.gridWidth
 
     invertedYRatio = gl.yRange / gl.gridWidth
     invertedXRatio = gl.xRange / gl.gridHeight
     zRatio = (gl.zRange / gl.gridHeight) * 16
+
+    if (!zRatio) {
+      zRatio = 1
+    }
 
     if (gl.minY !== Number.MIN_VALUE && Math.abs(gl.minY) !== 0) {
       // Negative numbers present in series
@@ -285,7 +297,7 @@ class CoreUtils {
       }
     })
 
-    return w.globals.seriesLog
+    return w.globals.invalidLogScale ? series : w.globals.seriesLog
   }
 
   getLogYRatios(yRatio) {
@@ -315,7 +327,7 @@ class CoreUtils {
       }
     })
 
-    return gl.yLogRatio
+    return gl.invalidLogScale ? yRatio.slice() : gl.yLogRatio
   }
 
   // Some config objects can be array - and we need to extend them correctly
