@@ -30,15 +30,14 @@ export default class Markers {
         }
       }
     } else {
-      w.globals.markers.size = w.config.series.map((s) => {
-        return w.config.markers.size
-      })
+      w.globals.markers.size = w.config.series.map((s) => w.config.markers.size)
     }
   }
 
   plotChartMarkers(pointsPos, seriesIndex, j) {
     let w = this.w
 
+    let i = seriesIndex
     let p = pointsPos
     let elPointsWrap = null
 
@@ -80,24 +79,26 @@ export default class Markers {
 
         if (shouldMarkerDraw) {
           if (Utils.isNumber(p.y[q])) {
-            PointClasses += ` w${(Math.random() + 1).toString(36).substring(4)}`
+            PointClasses += ` w${Utils.randomId()}`
           } else {
             PointClasses = 'apexcharts-nullpoint'
           }
 
-          let opts = this.getMarkerConfig(PointClasses, seriesIndex)
+          let opts = this.getMarkerConfig(
+            PointClasses,
+            seriesIndex,
+            dataPointIndex
+          )
 
-          // discrete markers is an option where user can specify a particular marker with different size and color
-          w.config.markers.discrete.map((marker) => {
-            if (
-              marker.seriesIndex === seriesIndex &&
-              marker.dataPointIndex === dataPointIndex
-            ) {
-              opts.pointStrokeColor = marker.strokeColor
-              opts.pointFillColor = marker.fillColor
-              opts.pSize = marker.size
+          if (w.config.series[i].data[j]) {
+            if (w.config.series[i].data[j].fillColor) {
+              opts.pointFillColor = w.config.series[i].data[j].fillColor
             }
-          })
+
+            if (w.config.series[i].data[j].strokeColor) {
+              opts.pointStrokeColor = w.config.series[i].data[j].strokeColor
+            }
+          }
 
           point = graphics.drawMarker(p.x[q], p.y[q], opts)
 
@@ -126,30 +127,58 @@ export default class Markers {
     return elPointsWrap
   }
 
-  getMarkerConfig(cssClass, seriesIndex) {
+  getMarkerConfig(cssClass, seriesIndex, dataPointIndex = null) {
     const w = this.w
     let pStyle = this.getMarkerStyle(seriesIndex)
+    let pSize = w.globals.markers.size[seriesIndex]
 
-    const pSize = w.globals.markers.size[seriesIndex]
+    const m = w.config.markers
+
+    // discrete markers is an option where user can specify a particular marker with different size and color
+
+    if (dataPointIndex !== null && m.discrete.length) {
+      m.discrete.map((marker) => {
+        if (
+          marker.seriesIndex === seriesIndex &&
+          marker.dataPointIndex === dataPointIndex
+        ) {
+          pStyle.pointStrokeColor = marker.strokeColor
+          pStyle.pointFillColor = marker.fillColor
+          pSize = marker.size
+        }
+      })
+    }
 
     return {
       pSize,
-      pRadius: w.config.markers.radius,
-      pWidth: w.config.markers.strokeWidth,
+      pRadius: m.radius,
+      pWidth:
+        m.strokeWidth instanceof Array
+          ? m.strokeWidth[seriesIndex]
+          : m.strokeWidth,
       pointStrokeColor: pStyle.pointStrokeColor,
       pointFillColor: pStyle.pointFillColor,
-      shape:
-        w.config.markers.shape instanceof Array
-          ? w.config.markers.shape[seriesIndex]
-          : w.config.markers.shape,
+      shape: m.shape instanceof Array ? m.shape[seriesIndex] : m.shape,
       class: cssClass,
-      pointStrokeOpacity: w.config.markers.strokeOpacity,
-      pointFillOpacity: w.config.markers.fillOpacity,
+      pointStrokeOpacity:
+        m.strokeOpacity instanceof Array
+          ? m.strokeOpacity[seriesIndex]
+          : m.strokeOpacity,
+      pointStrokeDashArray:
+        m.strokeDashArray instanceof Array
+          ? m.strokeDashArray[seriesIndex]
+          : m.strokeDashArray,
+      pointFillOpacity:
+        m.fillOpacity instanceof Array
+          ? m.fillOpacity[seriesIndex]
+          : m.fillOpacity,
       seriesIndex
     }
   }
 
   addEvents(circle) {
+    const w = this.w
+
     const graphics = new Graphics(this.ctx)
     circle.node.addEventListener(
       'mouseenter',
@@ -164,6 +193,9 @@ export default class Markers {
       'mousedown',
       graphics.pathMouseDown.bind(this.ctx, circle)
     )
+
+    circle.node.addEventListener('click', w.config.markers.onClick)
+    circle.node.addEventListener('dblclick', w.config.markers.onDblClick)
 
     circle.node.addEventListener(
       'touchstart',
