@@ -9,8 +9,14 @@ import Utils from '../utils/Utils'
 export default class Theme {
   constructor(ctx) {
     this.ctx = ctx
-    this.w = ctx.w
     this.colors = []
+    this.w = ctx.w
+    const w = this.w
+
+    this.isColorFn = false
+    this.isBarDistributed =
+      w.config.plotOptions.bar.distributed &&
+      (w.config.chart.type === 'bar' || w.config.chart.type === 'rangeBar')
   }
 
   init() {
@@ -21,19 +27,47 @@ export default class Theme {
     let w = this.w
     let utils = new Utils()
 
+    w.globals.dom.elWrap.classList.add(
+      `apexcharts-theme-${w.config.theme.mode}`
+    )
+
     if (w.config.colors === undefined) {
       w.globals.colors = this.predefined()
     } else {
       w.globals.colors = w.config.colors
+
+      // if user provided a function in colors, we need to eval here
+      if (
+        w.globals.axisCharts &&
+        w.config.chart.type !== 'bar' &&
+        Array.isArray(w.config.colors) &&
+        w.config.colors.length > 0 &&
+        w.config.colors.length === w.config.series.length
+        // colors & series length needs same
+      ) {
+        w.globals.colors = w.config.colors.map((c, i) => {
+          if (typeof c === 'function') {
+            this.isColorFn = true
+            return c({
+              value: w.globals.axisCharts
+                ? w.globals.series[i][0]
+                  ? w.globals.series[i][0]
+                  : 0
+                : w.globals.series[i],
+              seriesIndex: i,
+              dataPointIndex: i,
+              w
+            })
+          }
+          return c
+        })
+      }
     }
 
     if (w.config.theme.monochrome.enabled) {
       let monoArr = []
       let glsCnt = w.globals.series.length
-      if (
-        w.config.plotOptions.bar.distributed &&
-        w.config.chart.type === 'bar'
-      ) {
+      if (this.isBarDistributed) {
         glsCnt = w.globals.series[0].length * w.globals.series.length
       }
 
@@ -62,21 +96,15 @@ export default class Theme {
     // if user specfied less colors than no. of series, push the same colors again
     this.pushExtraColors(w.globals.colors)
 
-    // The Border colors
-    if (w.config.stroke.colors === undefined) {
-      w.globals.stroke.colors = defaultColors
-    } else {
-      w.globals.stroke.colors = w.config.stroke.colors
-    }
-    this.pushExtraColors(w.globals.stroke.colors)
-
-    // The FILL colors
-    if (w.config.fill.colors === undefined) {
-      w.globals.fill.colors = defaultColors
-    } else {
-      w.globals.fill.colors = w.config.fill.colors
-    }
-    this.pushExtraColors(w.globals.fill.colors)
+    const colorTypes = ['fill', 'stroke']
+    colorTypes.forEach((c) => {
+      if (w.config[c].colors === undefined) {
+        w.globals[c].colors = this.isColorFn ? w.config.colors : defaultColors
+      } else {
+        w.globals[c].colors = w.config[c].colors
+      }
+      this.pushExtraColors(w.globals[c].colors)
+    })
 
     if (w.config.dataLabels.style.colors === undefined) {
       w.globals.dataLabels.style.colors = defaultColors
@@ -86,7 +114,9 @@ export default class Theme {
     this.pushExtraColors(w.globals.dataLabels.style.colors, 50)
 
     if (w.config.plotOptions.radar.polygons.fill.colors === undefined) {
-      w.globals.radarPolygons.fill.colors = ['#fff']
+      w.globals.radarPolygons.fill.colors = [
+        w.config.theme.mode === 'dark' ? '#202D48' : '#fff'
+      ]
     } else {
       w.globals.radarPolygons.fill.colors =
         w.config.plotOptions.radar.polygons.fill.colors
@@ -113,8 +143,7 @@ export default class Theme {
 
     if (distributed === null) {
       distributed =
-        (w.config.chart.type === 'bar' &&
-          w.config.plotOptions.bar.distributed) ||
+        this.isBarDistributed ||
         (w.config.chart.type === 'heatmap' &&
           w.config.plotOptions.heatmap.colorScale.inverse)
     }
@@ -131,8 +160,31 @@ export default class Theme {
     }
   }
 
+  updateThemeOptions(options) {
+    options.chart = options.chart || {}
+    options.tooltip = options.tooltip || {}
+    const mode = options.theme.mode || 'light'
+    const palette = options.theme.palette
+      ? options.theme.palette
+      : mode === 'dark'
+      ? 'palette4'
+      : 'palette1'
+    const foreColor = options.chart.foreColor
+      ? options.chart.foreColor
+      : mode === 'dark'
+      ? '#f6f7f8'
+      : '#373d3f'
+
+    options.tooltip.theme = mode
+    options.chart.foreColor = foreColor
+    options.theme.palette = palette
+
+    return options
+  }
+
   predefined() {
     let palette = this.w.config.theme.palette
+
     // D6E3F8, FCEFEF, DCE0D9, A5978B, EDDDD4, D6E3F8, FEF5EF
     switch (palette) {
       case 'palette1':
@@ -145,7 +197,7 @@ export default class Theme {
         this.colors = ['#33b2df', '#546E7A', '#d4526e', '#13d8aa', '#A5978B']
         break
       case 'palette4':
-        this.colors = ['#546E7A', '#4ecdc4', '#c7f464', '#81D4FA', '#fd6a6a']
+        this.colors = ['#4ecdc4', '#c7f464', '#81D4FA', '#fd6a6a', '#546E7A']
         break
       case 'palette5':
         this.colors = ['#2b908f', '#f9a3a4', '#90ee7e', '#fa4443', '#69d2e7']
