@@ -1,5 +1,6 @@
 import Graphics from './Graphics'
 import Exports from './Exports'
+import Scales from './Scales'
 import Utils from './../utils/Utils'
 import icoPan from './../assets/ico-pan-hand.svg'
 import icoZoom from './../assets/ico-zoom-in.svg'
@@ -21,92 +22,91 @@ export default class Toolbar {
     this.w = ctx.w
 
     this.ev = this.w.config.chart.events
+    this.selectedClass = 'apexcharts-selected'
 
     this.localeValues = this.w.globals.locale.toolbar
   }
 
   createToolbar() {
     let w = this.w
-    const elToolbarWrap = document.createElement('div')
+
+    const createDiv = () => {
+      return document.createElement('div')
+    }
+    const elToolbarWrap = createDiv()
     elToolbarWrap.setAttribute('class', 'apexcharts-toolbar')
     w.globals.dom.elWrap.appendChild(elToolbarWrap)
 
-    this.elZoom = document.createElement('div')
-    this.elZoomIn = document.createElement('div')
-    this.elZoomOut = document.createElement('div')
-    this.elPan = document.createElement('div')
-    this.elSelection = document.createElement('div')
-    this.elZoomReset = document.createElement('div')
-    this.elMenuIcon = document.createElement('div')
-    this.elMenu = document.createElement('div')
+    this.elZoom = createDiv()
+    this.elZoomIn = createDiv()
+    this.elZoomOut = createDiv()
+    this.elPan = createDiv()
+    this.elSelection = createDiv()
+    this.elZoomReset = createDiv()
+    this.elMenuIcon = createDiv()
+    this.elMenu = createDiv()
     this.elCustomIcons = []
 
     this.t = w.config.chart.toolbar.tools
 
     if (Array.isArray(this.t.customIcons)) {
       for (let i = 0; i < this.t.customIcons.length; i++) {
-        this.elCustomIcons.push(document.createElement('div'))
+        this.elCustomIcons.push(createDiv())
       }
     }
 
-    this.elMenuItems = []
-
     let toolbarControls = []
 
-    if (this.t.zoomin && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoomIn,
-        icon: typeof this.t.zoomin === 'string' ? this.t.zoomin : icoZoomIn,
-        title: this.localeValues.zoomIn,
-        class: 'apexcharts-zoom-in-icon'
-      })
+    const appendZoomControl = (type, el, ico) => {
+      const tool = type.toLowerCase()
+      if (this.t[tool] && w.config.chart.zoom.enabled) {
+        toolbarControls.push({
+          el,
+          icon: typeof this.t[tool] === 'string' ? this.t[tool] : ico,
+          title: this.localeValues[type],
+          class: `apexcharts-${tool}-icon`
+        })
+      }
     }
 
-    if (this.t.zoomout && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoomOut,
-        icon: typeof this.t.zoomout === 'string' ? this.t.zoomout : icoZoomOut,
-        title: this.localeValues.zoomOut,
-        class: 'apexcharts-zoom-out-icon'
-      })
-    }
+    appendZoomControl('zoomIn', this.elZoomIn, icoZoomIn)
+    appendZoomControl('zoomOut', this.elZoomOut, icoZoomOut)
 
-    if (this.t.zoom && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoom,
-        icon: typeof this.t.zoom === 'string' ? this.t.zoom : icoZoom,
-        title: this.localeValues.selectionZoom,
-        class: w.globals.isTouchDevice ? 'hidden' : 'apexcharts-zoom-icon'
-      })
+    const zoomSelectionCtrls = (z) => {
+      if (this.t[z] && w.config.chart[z].enabled) {
+        toolbarControls.push({
+          el: z === 'zoom' ? this.elZoom : this.elSelection,
+          icon:
+            typeof this.t[z] === 'string'
+              ? this.t[z]
+              : z === 'zoom'
+              ? icoZoom
+              : icoSelect,
+          title: this.localeValues[
+            z === 'zoom' ? 'selectionZoom' : 'selection'
+          ],
+          class: w.globals.isTouchDevice
+            ? 'apexcharts-element-hidden'
+            : `apexcharts-${z}-icon`
+        })
+      }
     }
-
-    if (this.t.selection && w.config.chart.selection.enabled) {
-      toolbarControls.push({
-        el: this.elSelection,
-        icon:
-          typeof this.t.selection === 'string' ? this.t.selection : icoSelect,
-        title: this.localeValues.selection,
-        class: w.globals.isTouchDevice ? 'hidden' : 'apexcharts-selection-icon'
-      })
-    }
+    zoomSelectionCtrls('zoom')
+    zoomSelectionCtrls('selection')
 
     if (this.t.pan && w.config.chart.zoom.enabled) {
       toolbarControls.push({
         el: this.elPan,
         icon: typeof this.t.pan === 'string' ? this.t.pan : icoPan,
         title: this.localeValues.pan,
-        class: w.globals.isTouchDevice ? 'hidden' : 'apexcharts-pan-icon'
+        class: w.globals.isTouchDevice
+          ? 'apexcharts-element-hidden'
+          : 'apexcharts-pan-icon'
       })
     }
 
-    if (this.t.reset && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoomReset,
-        icon: typeof this.t.reset === 'string' ? this.t.reset : icoReset,
-        title: this.localeValues.reset,
-        class: 'apexcharts-reset-zoom-icon'
-      })
-    }
+    appendZoomControl('reset', this.elZoomReset, icoReset)
+
     if (this.t.download) {
       toolbarControls.push({
         el: this.elMenuIcon,
@@ -142,7 +142,22 @@ export default class Toolbar {
       elToolbarWrap.appendChild(toolbarControls[i].el)
     }
 
-    elToolbarWrap.appendChild(this.elMenu)
+    this._createHamburgerMenu(elToolbarWrap)
+
+    if (w.globals.zoomEnabled) {
+      this.elZoom.classList.add(this.selectedClass)
+    } else if (w.globals.panEnabled) {
+      this.elPan.classList.add(this.selectedClass)
+    } else if (w.globals.selectionEnabled) {
+      this.elSelection.classList.add(this.selectedClass)
+    }
+
+    this.addToolbarEventListeners()
+  }
+
+  _createHamburgerMenu(parent) {
+    this.elMenuItems = []
+    parent.appendChild(this.elMenu)
 
     Graphics.setAttrs(this.elMenu, {
       class: 'apexcharts-menu'
@@ -156,8 +171,17 @@ export default class Toolbar {
       {
         name: 'exportPNG',
         title: this.localeValues.exportToPNG
+      },
+      {
+        name: 'exportCSV',
+        title: this.localeValues.exportToCSV
       }
     ]
+
+    if (!this.w.globals.allSeriesHasEqualX) {
+      // if it is a multi series, and all series have variable x values, export CSV won't work
+      menuItems.splice(2, 1)
+    }
     for (let i = 0; i < menuItems.length; i++) {
       this.elMenuItems.push(document.createElement('div'))
       this.elMenuItems[i].innerHTML = menuItems[i].title
@@ -167,60 +191,51 @@ export default class Toolbar {
       })
       this.elMenu.appendChild(this.elMenuItems[i])
     }
-
-    if (w.globals.zoomEnabled) {
-      this.elZoom.classList.add('selected')
-    } else if (w.globals.panEnabled) {
-      this.elPan.classList.add('selected')
-    } else if (w.globals.selectionEnabled) {
-      this.elSelection.classList.add('selected')
-    }
-
-    this.addToolbarEventListeners()
   }
 
   addToolbarEventListeners() {
     this.elZoomReset.addEventListener('click', this.handleZoomReset.bind(this))
-    this.elSelection.addEventListener('click', this.toggleSelection.bind(this))
-    this.elZoom.addEventListener('click', this.toggleZooming.bind(this))
+    this.elSelection.addEventListener(
+      'click',
+      this.toggleZoomSelection.bind(this, 'selection')
+    )
+    this.elZoom.addEventListener(
+      'click',
+      this.toggleZoomSelection.bind(this, 'zoom')
+    )
     this.elZoomIn.addEventListener('click', this.handleZoomIn.bind(this))
     this.elZoomOut.addEventListener('click', this.handleZoomOut.bind(this))
     this.elPan.addEventListener('click', this.togglePanning.bind(this))
     this.elMenuIcon.addEventListener('click', this.toggleMenu.bind(this))
     this.elMenuItems.forEach((m) => {
       if (m.classList.contains('exportSVG')) {
-        m.addEventListener('click', this.downloadSVG.bind(this))
+        m.addEventListener('click', this.handleDownload.bind(this, 'svg'))
       } else if (m.classList.contains('exportPNG')) {
-        m.addEventListener('click', this.downloadPNG.bind(this))
+        m.addEventListener('click', this.handleDownload.bind(this, 'png'))
+      } else if (m.classList.contains('exportCSV')) {
+        m.addEventListener('click', this.handleDownload.bind(this, 'csv'))
       }
     })
     for (let i = 0; i < this.t.customIcons.length; i++) {
       this.elCustomIcons[i].addEventListener(
         'click',
-        this.t.customIcons[i].click
+        this.t.customIcons[i].click.bind(this, this.ctx, this.ctx.w)
       )
     }
   }
 
-  toggleSelection() {
+  toggleZoomSelection(type) {
     this.toggleOtherControls()
-    this.w.globals.selectionEnabled = !this.w.globals.selectionEnabled
 
-    if (!this.elSelection.classList.contains('selected')) {
-      this.elSelection.classList.add('selected')
+    let el = type === 'selection' ? this.elSelection : this.elZoom
+    let enabledType = type === 'selection' ? 'selectionEnabled' : 'zoomEnabled'
+
+    this.w.globals[enabledType] = !this.w.globals[enabledType]
+
+    if (!el.classList.contains(this.selectedClass)) {
+      el.classList.add(this.selectedClass)
     } else {
-      this.elSelection.classList.remove('selected')
-    }
-  }
-
-  toggleZooming() {
-    this.toggleOtherControls()
-    this.w.globals.zoomEnabled = !this.w.globals.zoomEnabled
-
-    if (!this.elZoom.classList.contains('selected')) {
-      this.elZoom.classList.add('selected')
-    } else {
-      this.elZoom.classList.remove('selected')
+      el.classList.remove(this.selectedClass)
     }
   }
 
@@ -239,26 +254,20 @@ export default class Toolbar {
     }
   }
 
-  enableZooming() {
+  enableZoomPanFromToolbar(type) {
     this.toggleOtherControls()
-    this.w.globals.zoomEnabled = true
-    if (this.elZoom) {
-      this.elZoom.classList.add('selected')
-    }
-    if (this.elPan) {
-      this.elPan.classList.remove('selected')
-    }
-  }
 
-  enablePanning() {
-    this.toggleOtherControls()
-    this.w.globals.panEnabled = true
+    type === 'pan'
+      ? (this.w.globals.panEnabled = true)
+      : (this.w.globals.zoomEnabled = true)
 
-    if (this.elPan) {
-      this.elPan.classList.add('selected')
+    const el = type === 'pan' ? this.elPan : this.elZoom
+    const el2 = type === 'pan' ? this.elZoom : this.elPan
+    if (el) {
+      el.classList.add(this.selectedClass)
     }
-    if (this.elZoom) {
-      this.elZoom.classList.remove('selected')
+    if (el2) {
+      el2.classList.remove(this.selectedClass)
     }
   }
 
@@ -266,10 +275,10 @@ export default class Toolbar {
     this.toggleOtherControls()
     this.w.globals.panEnabled = !this.w.globals.panEnabled
 
-    if (!this.elPan.classList.contains('selected')) {
-      this.elPan.classList.add('selected')
+    if (!this.elPan.classList.contains(this.selectedClass)) {
+      this.elPan.classList.add(this.selectedClass)
     } else {
-      this.elPan.classList.remove('selected')
+      this.elPan.classList.remove(this.selectedClass)
     }
   }
 
@@ -281,26 +290,25 @@ export default class Toolbar {
 
     this.getToolbarIconsReference()
 
-    if (this.elPan) {
-      this.elPan.classList.remove('selected')
-    }
-    if (this.elSelection) {
-      this.elSelection.classList.remove('selected')
-    }
-    if (this.elZoom) {
-      this.elZoom.classList.remove('selected')
-    }
+    const toggleEls = [this.elPan, this.elSelection, this.elZoom]
+    toggleEls.forEach((el) => {
+      if (el) {
+        el.classList.remove(this.selectedClass)
+      }
+    })
   }
 
   handleZoomIn() {
     const w = this.w
 
     const centerX = (w.globals.minX + w.globals.maxX) / 2
-    const newMinX = (w.globals.minX + centerX) / 2
-    const newMaxX = (w.globals.maxX + centerX) / 2
+    let newMinX = (w.globals.minX + centerX) / 2
+    let newMaxX = (w.globals.maxX + centerX) / 2
+
+    const newMinXMaxX = this._getNewMinXMaxX(newMinX, newMaxX)
 
     if (!w.globals.disableZoomIn) {
-      this.zoomUpdateOptions(newMinX, newMaxX)
+      this.zoomUpdateOptions(newMinXMaxX.minX, newMinXMaxX.maxX)
     }
   }
 
@@ -316,15 +324,39 @@ export default class Toolbar {
     }
 
     const centerX = (w.globals.minX + w.globals.maxX) / 2
-    const newMinX = w.globals.minX - (centerX - w.globals.minX)
-    const newMaxX = w.globals.maxX - (centerX - w.globals.maxX)
+    let newMinX = w.globals.minX - (centerX - w.globals.minX)
+    let newMaxX = w.globals.maxX - (centerX - w.globals.maxX)
+
+    const newMinXMaxX = this._getNewMinXMaxX(newMinX, newMaxX)
 
     if (!w.globals.disableZoomOut) {
-      this.zoomUpdateOptions(newMinX, newMaxX)
+      this.zoomUpdateOptions(newMinXMaxX.minX, newMinXMaxX.maxX)
+    }
+  }
+
+  _getNewMinXMaxX(newMinX, newMaxX) {
+    const shouldFloor = this.w.config.xaxis.convertedCatToNumeric
+    return {
+      minX: shouldFloor ? Math.floor(newMinX) : newMinX,
+      maxX: shouldFloor ? Math.floor(newMaxX) : newMaxX
     }
   }
 
   zoomUpdateOptions(newMinX, newMaxX) {
+    const w = this.w
+
+    if (w.config.xaxis.convertedCatToNumeric) {
+      // in category charts, avoid zooming out beyond min and max
+      if (newMinX < 1) {
+        newMinX = 1
+        newMaxX = w.globals.dataPoints
+      }
+
+      if (newMaxX - newMinX < 2) {
+        return
+      }
+    }
+
     let xaxis = {
       min: newMinX,
       max: newMaxX
@@ -335,17 +367,33 @@ export default class Toolbar {
       xaxis = beforeZoomRange.xaxis
     }
 
+    let options = {
+      xaxis
+    }
+
+    let yaxis = Utils.clone(w.globals.initialConfig.yaxis)
+    if (w.config.chart.zoom.autoScaleYaxis) {
+      const scale = new Scales(this.ctx)
+      yaxis = scale.autoScaleY(this.ctx, yaxis, {
+        xaxis
+      })
+    }
+
+    if (!w.config.chart.group) {
+      // if chart in a group, prevent yaxis update here
+      // fix issue #650
+      options.yaxis = yaxis
+    }
+
     this.w.globals.zoomed = true
 
-    this.ctx._updateOptions(
-      {
-        xaxis
-      },
+    this.ctx.updateHelpers._updateOptions(
+      options,
       false,
       this.w.config.chart.animations.dynamicAnimation.enabled
     )
 
-    this.zoomCallback(xaxis)
+    this.zoomCallback(xaxis, yaxis)
   }
 
   zoomCallback(xaxis, yaxis) {
@@ -364,23 +412,29 @@ export default class Toolbar {
   }
 
   toggleMenu() {
-    if (this.elMenu.classList.contains('open')) {
-      this.elMenu.classList.remove('open')
-    } else {
-      this.elMenu.classList.add('open')
+    window.setTimeout(() => {
+      if (this.elMenu.classList.contains('apexcharts-menu-open')) {
+        this.elMenu.classList.remove('apexcharts-menu-open')
+      } else {
+        this.elMenu.classList.add('apexcharts-menu-open')
+      }
+    }, 0)
+  }
+
+  handleDownload(type) {
+    const w = this.w
+    const exprt = new Exports(this.ctx)
+    switch (type) {
+      case 'svg':
+        exprt.exportToSVG(this.ctx)
+        break
+      case 'png':
+        exprt.exportToPng(this.ctx)
+        break
+      case 'csv':
+        exprt.exportToCSV({ series: w.config.series })
+        break
     }
-  }
-
-  downloadPNG() {
-    const downloadPNG = new Exports(this.ctx)
-    downloadPNG.exportToPng(this.ctx)
-    this.toggleMenu()
-  }
-
-  downloadSVG() {
-    const downloadSVG = new Exports(this.ctx)
-    downloadSVG.exportToSVG()
-    this.toggleMenu()
   }
 
   handleZoomReset(e) {
@@ -390,10 +444,10 @@ export default class Toolbar {
       let w = ch.w
 
       if (
-        w.globals.minX !== w.globals.initialminX &&
-        w.globals.maxX !== w.globals.initialmaxX
+        w.globals.minX !== w.globals.initialMinX ||
+        w.globals.maxX !== w.globals.initialMaxX
       ) {
-        ch.revertDefaultAxisMinMax()
+        ch.updateHelpers.revertDefaultAxisMinMax()
 
         if (typeof w.config.chart.events.zoomed === 'function') {
           this.zoomCallback({
@@ -404,7 +458,7 @@ export default class Toolbar {
 
         w.globals.zoomed = false
 
-        ch._updateSeries(
+        ch.updateHelpers._updateSeries(
           w.globals.initialSeries,
           w.config.chart.animations.dynamicAnimation.enabled
         )
@@ -413,22 +467,6 @@ export default class Toolbar {
   }
 
   destroy() {
-    if (this.elZoomReset) {
-      this.elZoomReset.removeEventListener(
-        'click',
-        this.handleZoomReset.bind(this)
-      )
-      this.elSelection.removeEventListener(
-        'click',
-        this.toggleSelection.bind(this)
-      )
-      this.elZoom.removeEventListener('click', this.toggleZooming.bind(this))
-      this.elZoomIn.removeEventListener('click', this.handleZoomIn.bind(this))
-      this.elZoomOut.removeEventListener('click', this.handleZoomOut.bind(this))
-      this.elPan.removeEventListener('click', this.togglePanning.bind(this))
-      this.elMenuIcon.removeEventListener('click', this.toggleMenu.bind(this))
-    }
-
     this.elZoom = null
     this.elZoomIn = null
     this.elZoomOut = null
