@@ -1,6 +1,5 @@
 import Defaults from './Defaults'
 import Utils from './../../utils/Utils'
-import CoreUtils from '../CoreUtils'
 import Options from './Options'
 
 /**
@@ -13,7 +12,7 @@ export default class Config {
     this.opts = opts
   }
 
-  init() {
+  init({ responsiveOverride }) {
     let opts = this.opts
     let options = new Options()
     let defaults = new Defaults(opts)
@@ -35,8 +34,6 @@ export default class Config {
       )
     }
 
-    opts.series = this.checkEmptySeries(opts.series)
-
     opts = this.extendYAxis(opts)
     opts = this.extendAnnotations(opts)
 
@@ -44,45 +41,26 @@ export default class Config {
     let newDefaults = {}
     if (opts && typeof opts === 'object') {
       let chartDefaults = {}
-      switch (this.chartType) {
-        case 'line':
-          chartDefaults = defaults.line()
-          break
-        case 'area':
-          chartDefaults = defaults.area()
-          break
-        case 'bar':
-          chartDefaults = defaults.bar()
-          break
-        case 'candlestick':
-          chartDefaults = defaults.candlestick()
-          break
-        case 'histogram':
-          chartDefaults = defaults.bar()
-          break
-        case 'bubble':
-          chartDefaults = defaults.bubble()
-          break
-        case 'scatter':
-          chartDefaults = defaults.scatter()
-          break
-        case 'heatmap':
-          chartDefaults = defaults.heatmap()
-          break
-        case 'pie':
-          chartDefaults = defaults.pie()
-          break
-        case 'donut':
-          chartDefaults = defaults.donut()
-          break
-        case 'radar':
-          chartDefaults = defaults.radar()
-          break
-        case 'radialBar':
-          chartDefaults = defaults.radialBar()
-          break
-        default:
-          chartDefaults = defaults.line()
+      const chartTypes = [
+        'line',
+        'area',
+        'bar',
+        'candlestick',
+        'rangeBar',
+        'histogram',
+        'bubble',
+        'scatter',
+        'heatmap',
+        'pie',
+        'donut',
+        'radar',
+        'radialBar'
+      ]
+
+      if (chartTypes.indexOf(opts.chart.type) !== -1) {
+        chartDefaults = defaults[opts.chart.type]()
+      } else {
+        chartDefaults = defaults.line()
       }
 
       if (opts.chart.brush && opts.chart.brush.enabled) {
@@ -90,22 +68,23 @@ export default class Config {
       }
 
       if (opts.chart.stacked && opts.chart.stackType === '100%') {
-        defaults.stacked100()
+        opts = defaults.stacked100(opts)
       }
+
+      // If user has specified a dark theme, make the tooltip dark too
+      this.checkForDarkTheme(window.Apex) // check global window Apex options
+      this.checkForDarkTheme(opts) // check locally passed options
 
       opts.xaxis = opts.xaxis || window.Apex.xaxis || {}
 
-      const combo = CoreUtils.checkComboSeries(opts.series)
-      if (
-        (opts.chart.type === 'line' ||
-          opts.chart.type === 'area' ||
-          opts.chart.type === 'scatter') &&
-        !combo.comboChartsHasBars &&
-        opts.xaxis.type !== 'datetime' &&
-        opts.xaxis.tickPlacement !== 'between'
-      ) {
-        opts = Defaults.convertCatToNumeric(opts)
+      // an important boolean needs to be set here
+      // otherwise all the charts will have this flag set to true window.Apex.xaxis is set globally
+      if (!responsiveOverride) {
+        opts.xaxis.convertedCatToNumeric = false
       }
+
+      opts = this.checkForCatToNumericXAxis(this.chartType, chartDefaults, opts)
+
       if (
         (opts.chart.sparkline && opts.chart.sparkline.enabled) ||
         (window.Apex.chart &&
@@ -132,9 +111,48 @@ export default class Config {
     return config
   }
 
+  checkForCatToNumericXAxis(chartType, chartDefaults, opts) {
+    let defaults = new Defaults(opts)
+
+    const isBarHorizontal =
+      chartType === 'bar' &&
+      opts.plotOptions &&
+      opts.plotOptions.bar &&
+      opts.plotOptions.bar.horizontal
+
+    const unsupportedZoom =
+      chartType === 'pie' ||
+      chartType === 'donut' ||
+      chartType === 'radar' ||
+      chartType === 'radialBar' ||
+      chartType === 'heatmap'
+
+    const notNumericXAxis =
+      opts.xaxis.type !== 'datetime' && opts.xaxis.type !== 'numeric'
+
+    let tickPlacement = opts.xaxis.tickPlacement
+      ? opts.xaxis.tickPlacement
+      : chartDefaults.xaxis && chartDefaults.xaxis.tickPlacement
+    if (
+      !isBarHorizontal &&
+      !unsupportedZoom &&
+      notNumericXAxis &&
+      tickPlacement !== 'between'
+    ) {
+      opts = defaults.convertCatToNumeric(opts)
+    }
+
+    return opts
+  }
+
   extendYAxis(opts) {
     let options = new Options()
-    if (typeof opts.yaxis === 'undefined') {
+
+    if (
+      typeof opts.yaxis === 'undefined' ||
+      !opts.yaxis ||
+      (Array.isArray(opts.yaxis) && opts.yaxis.length === 0)
+    ) {
       opts.yaxis = {}
     }
 
@@ -154,6 +172,45 @@ export default class Config {
       opts.yaxis = [Utils.extend(options.yAxis, opts.yaxis)]
     } else {
       opts.yaxis = Utils.extendArray(opts.yaxis, options.yAxis)
+    }
+
+    let isLogY = false
+    opts.yaxis.forEach((y) => {
+      if (y.logarithmic) {
+        isLogY = true
+      }
+    })
+
+    // A logarithmic chart works correctly when each series has a corresponding y-axis
+    // If this is not the case, we manually create yaxis for multi-series log chart
+    if (
+      isLogY &&
+      opts.series.length !== opts.yaxis.length &&
+      opts.series.length
+    ) {
+      opts.yaxis = opts.series.map((s, i) => {
+        if (!s.name) {
+          opts.series[i].name = `series-${i + 1}`
+        }
+        if (opts.yaxis[i]) {
+          opts.yaxis[i].seriesName = opts.series[i].name
+          return opts.yaxis[i]
+        } else {
+          const newYaxis = Utils.extend(options.yAxis, opts.yaxis[0])
+          newYaxis.show = false
+          return newYaxis
+        }
+      })
+    }
+
+    if (
+      isLogY &&
+      opts.series.length > 1 &&
+      opts.series.length !== opts.yaxis.length
+    ) {
+      console.warn(
+        'A multi-series logarithmic chart should have equal number of series and y-axes. Please make sure to equalize both.'
+      )
     }
     return opts
   }
@@ -176,6 +233,7 @@ export default class Config {
 
   extendYAxisAnnotations(opts) {
     let options = new Options()
+
     opts.annotations.yaxis = Utils.extendArray(
       typeof opts.annotations.yaxis !== 'undefined'
         ? opts.annotations.yaxis
@@ -187,6 +245,7 @@ export default class Config {
 
   extendXAxisAnnotations(opts) {
     let options = new Options()
+
     opts.annotations.xaxis = Utils.extendArray(
       typeof opts.annotations.xaxis !== 'undefined'
         ? opts.annotations.xaxis
@@ -197,6 +256,7 @@ export default class Config {
   }
   extendPointAnnotations(opts) {
     let options = new Options()
+
     opts.annotations.points = Utils.extendArray(
       typeof opts.annotations.points !== 'undefined'
         ? opts.annotations.points
@@ -206,15 +266,23 @@ export default class Config {
     return opts
   }
 
-  checkEmptySeries(ser) {
-    if (ser.length === 0) {
-      return [
-        {
-          data: []
-        }
-      ]
+  checkForDarkTheme(opts) {
+    if (opts.theme && opts.theme.mode === 'dark') {
+      if (!opts.tooltip) {
+        opts.tooltip = {}
+      }
+      if (opts.tooltip.theme !== 'light') {
+        opts.tooltip.theme = 'dark'
+      }
+
+      if (!opts.chart.foreColor) {
+        opts.chart.foreColor = '#f6f7f8'
+      }
+
+      if (!opts.theme.palette) {
+        opts.theme.palette = 'palette4'
+      }
     }
-    return ser
   }
 
   handleUserInputErrors(opts) {
@@ -226,20 +294,10 @@ export default class Config {
       )
     }
 
-    if (config.chart.scroller) {
-      console.warn(
-        'Scroller has been deprecated since v2.0.0. Please remove the configuration for chart.scroller'
-      )
-    }
-
-    if (config.chart.type === 'bar' && config.plotOptions.bar.horizontal) {
-      // No time series for horizontal bars
-      if (config.xaxis.type === 'datetime') {
-        throw new Error(
-          'Timelines on bars are not supported yet. Switch to column chart by setting plotOptions.bar.horizontal=false'
-        )
-      }
-
+    if (
+      (config.chart.type === 'bar' || config.chart.type === 'rangeBar') &&
+      config.plotOptions.bar.horizontal
+    ) {
       // No multiple yaxis for bars
       if (config.yaxis.length > 1) {
         throw new Error(
@@ -247,6 +305,7 @@ export default class Config {
         )
       }
 
+      // if yaxis is reversed in horizontal bar chart, you should draw the y-axis on right side
       if (config.yaxis[0].reversed) {
         config.yaxis[0].opposite = true
       }
@@ -256,7 +315,7 @@ export default class Config {
       config.chart.zoom.enabled = false // no zooming for horz bars
     }
 
-    if (config.chart.type === 'bar') {
+    if (config.chart.type === 'bar' || config.chart.type === 'rangeBar') {
       if (config.tooltip.shared) {
         if (
           config.xaxis.crosshairs.width === 'barWidth' &&
@@ -269,10 +328,12 @@ export default class Config {
         }
         if (config.plotOptions.bar.horizontal) {
           config.states.hover.type = 'none'
+          config.tooltip.shared = false
         }
         if (!config.tooltip.followCursor) {
           console.warn(
-            'followCursor option in shared columns cannot be turned off.'
+            'followCursor option in shared columns cannot be turned off. Please set %ctooltip.followCursor: true',
+            'color: blue;'
           )
           config.tooltip.followCursor = true
         }
