@@ -35,6 +35,9 @@ class Intersect {
         shared: false
       })
 
+      w.globals.capturedSeriesIndex = i
+      w.globals.capturedDataPointIndex = j
+
       x = cx + ttCtx.tooltipRect.ttWidth / 2 + width
       y = cy + ttCtx.tooltipRect.ttHeight / 2 - height / 2
 
@@ -64,20 +67,21 @@ class Intersect {
     let i
     let j
     if (e.target.classList.contains('apexcharts-marker')) {
-      let cx = parseInt(opt.paths.getAttribute('cx'))
-      let cy = parseInt(opt.paths.getAttribute('cy'))
+      let cx = parseInt(opt.paths.getAttribute('cx'), 10)
+      let cy = parseInt(opt.paths.getAttribute('cy'), 10)
       let val = parseFloat(opt.paths.getAttribute('val'))
 
-      j = parseInt(opt.paths.getAttribute('rel'))
+      j = parseInt(opt.paths.getAttribute('rel'), 10)
       i =
         parseInt(
-          opt.paths.parentNode.parentNode.parentNode.getAttribute('rel')
+          opt.paths.parentNode.parentNode.parentNode.getAttribute('rel'),
+          10
         ) - 1
 
       if (ttCtx.intersect) {
         const el = Utils.findAncestor(opt.paths, 'apexcharts-series')
         if (el) {
-          i = parseInt(el.getAttribute('data:realIndex'))
+          i = parseInt(el.getAttribute('data:realIndex'), 10)
         }
       }
 
@@ -85,23 +89,29 @@ class Intersect {
         ttItems: opt.ttItems,
         i,
         j,
-        shared: ttCtx.intersect ? false : w.config.tooltip.shared
+        shared: ttCtx.showOnIntersect ? false : w.config.tooltip.shared
       })
 
-      ttCtx.marker.enlargeCurrentPoint(j, opt.paths)
+      if (e.type === 'mouseup') {
+        ttCtx.markerClick(e, i, j)
+      }
+
+      w.globals.capturedSeriesIndex = i
+      w.globals.capturedDataPointIndex = j
 
       x = cx
-      y = cy - ttCtx.tooltipRect.ttHeight * 1.4
+      y = cy + w.globals.translateY - ttCtx.tooltipRect.ttHeight * 1.4
 
       if (ttCtx.w.config.tooltip.followCursor) {
         const elGrid = ttCtx.getElGrid()
         const seriesBound = elGrid.getBoundingClientRect()
-        y = ttCtx.e.clientY - seriesBound.top
+        y = ttCtx.e.clientY + w.globals.translateY - seriesBound.top
       }
 
       if (val < 0) {
         y = cy
       }
+      ttCtx.marker.enlargeCurrentPoint(j, opt.paths, x, y)
     }
 
     return {
@@ -119,7 +129,6 @@ class Intersect {
     let bx = 0
     let x = 0
     let y = 0
-    // let bW = 0
     let i = 0
     let strokeWidth
     let barXY = this.getBarTooltipXY({
@@ -130,8 +139,11 @@ class Intersect {
     let barHeight = barXY.barHeight
     let j = barXY.j
 
+    w.globals.capturedSeriesIndex = i
+    w.globals.capturedDataPointIndex = j
+
     if (
-      (ttCtx.isBarHorizontal && ttCtx.hasBars()) ||
+      (w.globals.isBarHorizontal && ttCtx.tooltipUtil.hasBars()) ||
       !w.config.tooltip.shared
     ) {
       x = barXY.x
@@ -139,7 +151,6 @@ class Intersect {
       strokeWidth = Array.isArray(w.config.stroke.width)
         ? w.config.stroke.width[i]
         : w.config.stroke.width
-      // bW = barXY.barWidth
       bx = x
     } else {
       if (!w.globals.comboCharts && !w.config.tooltip.shared) {
@@ -150,13 +161,14 @@ class Intersect {
     // y is NaN, make it touch the bottom of grid area
     if (isNaN(y)) {
       y = w.globals.svgHeight - ttCtx.tooltipRect.ttHeight
+    } else if (y < 0) {
+      y = 0
     }
 
-    // x exceeds gridWidth
     if (x + ttCtx.tooltipRect.ttWidth > w.globals.gridWidth) {
       x = x - ttCtx.tooltipRect.ttWidth
     } else if (x < 0) {
-      x = x + ttCtx.tooltipRect.ttWidth
+      x = 0
     }
 
     if (ttCtx.w.config.tooltip.followCursor) {
@@ -171,7 +183,7 @@ class Intersect {
     }
 
     if (!w.config.tooltip.shared) {
-      if (w.globals.comboChartsHasBars) {
+      if (w.globals.comboBarCount > 0) {
         ttCtx.tooltipPosition.moveXCrosshairs(bx + strokeWidth / 2)
       } else {
         ttCtx.tooltipPosition.moveXCrosshairs(bx)
@@ -181,22 +193,30 @@ class Intersect {
     // move tooltip here
     if (
       !ttCtx.fixedTooltip &&
-      (!w.config.tooltip.shared || (ttCtx.isBarHorizontal && ttCtx.hasBars()))
+      (!w.config.tooltip.shared ||
+        (w.globals.isBarHorizontal && ttCtx.tooltipUtil.hasBars()))
     ) {
+      const isReversed = w.globals.isMultipleYAxis
+        ? w.config.yaxis[seriesIndex] && w.config.yaxis[seriesIndex].reversed
+        : w.config.yaxis[0].reversed
+
       if (isReversed) {
-        x = w.globals.gridWidth - x
+        x = x - ttCtx.tooltipRect.ttWidth
+        if (x < 0) {
+          x = 0
+        }
       }
       tooltipEl.style.left = x + w.globals.translateX + 'px'
 
       const seriesIndex = parseInt(
-        opt.paths.parentNode.getAttribute('data:realIndex')
+        opt.paths.parentNode.getAttribute('data:realIndex'),
+        10
       )
 
-      const isReversed = w.globals.isMultipleYAxis
-        ? w.config.yaxis[seriesIndex].reversed
-        : w.config.yaxis[0].reversed
-
-      if (isReversed && !(ttCtx.isBarHorizontal && ttCtx.hasBars())) {
+      if (
+        isReversed &&
+        !(w.globals.isBarHorizontal && ttCtx.tooltipUtil.hasBars())
+      ) {
         y = y + barHeight - (w.globals.series[i][j] < 0 ? barHeight : 0) * 2
       }
       if (ttCtx.tooltipRect.ttHeight + y > w.globals.gridHeight) {
@@ -226,7 +246,8 @@ class Intersect {
 
     if (
       cl.contains('apexcharts-bar-area') ||
-      cl.contains('apexcharts-candlestick-area')
+      cl.contains('apexcharts-candlestick-area') ||
+      cl.contains('apexcharts-rangebar-area')
     ) {
       let bar = e.target
       let barRect = bar.getBoundingClientRect()
@@ -237,16 +258,19 @@ class Intersect {
       barHeight = barRect.height
       let bw = barRect.width
 
-      let cx = parseInt(bar.getAttribute('cx'))
-      let cy = parseInt(bar.getAttribute('cy'))
+      let cx = parseInt(bar.getAttribute('cx'), 10)
+      let cy = parseInt(bar.getAttribute('cy'), 10)
       barWidth = parseFloat(bar.getAttribute('barWidth'))
       const clientX = e.type === 'touchmove' ? e.touches[0].clientX : e.clientX
 
-      j = parseInt(bar.getAttribute('j'))
-      i = parseInt(bar.parentNode.getAttribute('rel')) - 1
+      j = parseInt(bar.getAttribute('j'), 10)
+      i = parseInt(bar.parentNode.getAttribute('rel'), 10) - 1
+
+      let y1 = bar.getAttribute('data-range-y1')
+      let y2 = bar.getAttribute('data-range-y2')
 
       if (w.globals.comboCharts) {
-        i = parseInt(bar.parentNode.getAttribute('data:realIndex'))
+        i = parseInt(bar.parentNode.getAttribute('data:realIndex'), 10)
       }
 
       // if (w.config.tooltip.shared) {
@@ -261,11 +285,13 @@ class Intersect {
         ttItems: opt.ttItems,
         i,
         j,
+        y1: y1 ? parseInt(y1, 10) : null,
+        y2: y2 ? parseInt(y2, 10) : null,
         shared: ttCtx.showOnIntersect ? false : w.config.tooltip.shared
       })
 
       if (w.config.tooltip.followCursor) {
-        if (w.config.plotOptions.bar.horizontal) {
+        if (w.globals.isBarHorizontal) {
           x = clientX - seriesBound.left + 15
           y =
             cy -
@@ -281,7 +307,7 @@ class Intersect {
           y = e.clientY - seriesBound.top - ttCtx.tooltipRect.ttHeight / 2 - 15
         }
       } else {
-        if (w.config.plotOptions.bar.horizontal) {
+        if (w.globals.isBarHorizontal) {
           x = cx
           if (x < ttCtx.xyRatios.baseLineInvertedY) {
             x = cx - ttCtx.tooltipRect.ttWidth
