@@ -33,17 +33,12 @@ export default class Position {
       x = (w.globals.gridWidth / tickAmount) * j
     }
 
-    if (
-      w.config.xaxis.crosshairs.width === 'tickWidth' ||
-      w.config.xaxis.crosshairs.width === 'barWidth'
-    ) {
-      if (x + ttCtx.xcrosshairsWidth > w.globals.gridWidth) {
-        x = w.globals.gridWidth - ttCtx.xcrosshairsWidth
-      }
-    } else {
-      if (j !== null) {
-        x = x + w.globals.gridWidth / tickAmount / 2
-      }
+    if (xcrosshairs !== null) {
+      xcrosshairs.setAttribute('x', x)
+      xcrosshairs.setAttribute('x1', x)
+      xcrosshairs.setAttribute('x2', x)
+      xcrosshairs.setAttribute('y2', w.globals.gridHeight)
+      xcrosshairs.classList.add('apexcharts-active')
     }
 
     if (x < 0) {
@@ -52,14 +47,6 @@ export default class Position {
 
     if (x > w.globals.gridWidth) {
       x = w.globals.gridWidth
-    }
-
-    if (xcrosshairs !== null) {
-      xcrosshairs.setAttribute('x', x)
-      xcrosshairs.setAttribute('x1', x)
-      xcrosshairs.setAttribute('x2', x)
-      xcrosshairs.setAttribute('y2', w.globals.gridHeight)
-      xcrosshairs.classList.add('active')
     }
 
     if (ttCtx.blxaxisTooltip) {
@@ -106,7 +93,7 @@ export default class Position {
     const ttCtx = this.ttCtx
 
     if (ttCtx.xaxisTooltip !== null) {
-      ttCtx.xaxisTooltip.classList.add('active')
+      ttCtx.xaxisTooltip.classList.add('apexcharts-active')
 
       let cy =
         ttCtx.xaxisOffY +
@@ -145,7 +132,8 @@ export default class Position {
     }
 
     const ycrosshairsHiddenRectY1 = parseInt(
-      ttCtx.ycrosshairsHidden.getAttribute('y1')
+      ttCtx.ycrosshairsHidden.getAttribute('y1'),
+      10
     )
     let cy = w.globals.translateY + ycrosshairsHiddenRectY1
 
@@ -160,12 +148,12 @@ export default class Position {
     cy = cy - yAxisTTHeight / 2
 
     if (w.globals.ignoreYAxisIndexes.indexOf(index) === -1) {
-      ttCtx.yaxisTTEls[index].classList.add('active')
+      ttCtx.yaxisTTEls[index].classList.add('apexcharts-active')
       ttCtx.yaxisTTEls[index].style.top = cy + 'px'
       ttCtx.yaxisTTEls[index].style.left =
         cx + w.config.yaxis[index].tooltip.offsetX + 'px'
     } else {
-      ttCtx.yaxisTTEls[index].classList.remove('active')
+      ttCtx.yaxisTTEls[index].classList.remove('apexcharts-active')
     }
   }
 
@@ -183,10 +171,10 @@ export default class Position {
     const tooltipEl = ttCtx.getElTooltip()
     let tooltipRect = ttCtx.tooltipRect
 
-    let pointR = r !== null ? parseInt(r) : 1
+    let pointR = r !== null ? parseFloat(r) : 1
 
-    let x = parseInt(cx) + pointR + 5
-    let y = parseInt(cy) + pointR / 2 // - tooltipRect.ttHeight / 2
+    let x = parseFloat(cx) + pointR + 5
+    let y = parseFloat(cy) + pointR / 2 // - tooltipRect.ttHeight / 2
 
     if (x > w.globals.gridWidth / 2) {
       x = x - tooltipRect.ttWidth - pointR - 15
@@ -203,7 +191,11 @@ export default class Position {
     if (w.config.tooltip.followCursor) {
       const elGrid = ttCtx.getElGrid()
       const seriesBound = elGrid.getBoundingClientRect()
-      y = ttCtx.e.clientY - seriesBound.top - tooltipRect.ttHeight / 2
+      y =
+        ttCtx.e.clientY +
+        w.globals.translateY -
+        seriesBound.top -
+        tooltipRect.ttHeight / 2
     }
 
     const newPositions = this.positionChecks(tooltipRect, x, y)
@@ -243,7 +235,7 @@ export default class Position {
         ` .apexcharts-series[data\\:realIndex='${i}'] .apexcharts-marker`
       )
       for (let p = 0; p < allPoints.length; p++) {
-        if (parseInt(allPoints[p].getAttribute('rel')) === j) {
+        if (parseInt(allPoints[p].getAttribute('rel'), 10) === j) {
           ttCtx.marker.resetPointsSize()
           ttCtx.marker.enlargeCurrentPoint(j, allPoints[p])
         }
@@ -264,12 +256,12 @@ export default class Position {
 
     let pointsArr = w.globals.pointsArray
 
-    let hoverSize = w.config.markers.hover.size
+    let hoverSize = ttCtx.tooltipUtil.getHoverMarkerSize(capturedSeries)
 
-    if (hoverSize === undefined) {
-      hoverSize =
-        w.globals.markers.size[capturedSeries] +
-        w.config.markers.hover.sizeOffset
+    const serType = w.config.series[capturedSeries].type
+    if (serType && (serType === 'column' || serType === 'candlestick')) {
+      // fix error mentioned in #811
+      return
     }
 
     cx = pointsArr[capturedSeries][j][0]
@@ -279,10 +271,13 @@ export default class Position {
       `.apexcharts-series[data\\:realIndex='${capturedSeries}'] .apexcharts-series-markers circle`
     )
 
-    point.setAttribute('r', hoverSize)
+    if (point) {
+      point.setAttribute('r', hoverSize)
 
-    point.setAttribute('cx', cx)
-    point.setAttribute('cy', cy)
+      point.setAttribute('cx', cx)
+      point.setAttribute('cy', cy)
+    }
+
     // point.style.opacity = w.config.markers.hover.opacity
 
     this.moveXCrosshairs(cx)
@@ -304,13 +299,9 @@ export default class Position {
     let pointsArr = w.globals.pointsArray
 
     let series = new Series(this.ctx)
-    activeSeries = series.getActiveSeriesIndex()
+    activeSeries = series.getActiveConfigSeriesIndex(true)
 
-    let hoverSize = w.config.markers.hover.size
-    if (hoverSize === undefined) {
-      hoverSize =
-        w.globals.markers.size[activeSeries] + w.config.markers.hover.sizeOffset
-    }
+    let hoverSize = ttCtx.tooltipUtil.getHoverMarkerSize(activeSeries)
 
     if (pointsArr[activeSeries]) {
       cx = pointsArr[activeSeries][j][0]
@@ -318,7 +309,7 @@ export default class Position {
     }
 
     let points = null
-    const allPoints = ttCtx.getAllMarkers()
+    const allPoints = ttCtx.tooltipUtil.getAllMarkers()
 
     if (allPoints !== null) {
       points = allPoints
@@ -329,19 +320,28 @@ export default class Position {
     }
 
     if (points !== null) {
-      for (let p = 0; p < points.length; p++) {
+      for (let p = 0; p < w.globals.series.length; p++) {
         let pointArr = pointsArr[p]
 
+        if (w.globals.comboCharts) {
+          // in a combo chart, if column charts are present, markers will not match with the number of series, hence this patch to push a null value in points array
+          if (typeof pointArr === 'undefined') {
+            // nodelist to array
+            points = [...points]
+            points.splice(p, 0, null)
+          }
+        }
         if (pointArr && pointArr.length) {
           let pcy = pointsArr[p][j][1]
           points[p].setAttribute('cx', cx)
           let realIndex = parseInt(
             points[p].parentNode.parentNode.parentNode.getAttribute(
               'data:realIndex'
-            )
+            ),
+            10
           )
 
-          if (pcy !== null) {
+          if (pcy !== null && !isNaN(pcy)) {
             points[realIndex] && points[realIndex].setAttribute('r', hoverSize)
             points[realIndex] && points[realIndex].setAttribute('cy', pcy)
           } else {
@@ -363,8 +363,17 @@ export default class Position {
     const w = this.w
     const ttCtx = this.ttCtx
 
+    let barLen = w.globals.columnSeries
+      ? w.globals.columnSeries.length
+      : w.globals.series.length
+
+    const i =
+      barLen >= 2 && barLen % 2 === 0
+        ? Math.floor(barLen / 2)
+        : Math.floor(barLen / 2) + 1
+
     let jBar = w.globals.dom.baseEl.querySelector(
-      `.apexcharts-bar-series .apexcharts-series[rel='1'] path[j='${j}'], .apexcharts-candlestick-series .apexcharts-series[rel='1'] path[j='${j}']`
+      `.apexcharts-bar-series .apexcharts-series[rel='${i}'] path[j='${j}'], .apexcharts-candlestick-series .apexcharts-series[rel='${i}'] path[j='${j}'], .apexcharts-rangebar-series .apexcharts-series[rel='${i}'] path[j='${j}']`
     )
 
     let bcx = jBar ? parseFloat(jBar.getAttribute('cx')) : 0
@@ -372,7 +381,7 @@ export default class Position {
     let bw = jBar ? parseFloat(jBar.getAttribute('barWidth')) : 0
 
     if (w.globals.isXNumeric) {
-      bcx = bcx - bw / 2
+      bcx = bcx - (barLen % 2 !== 0 ? bw / 2 : 0)
     } else {
       bcx = ttCtx.xAxisTicksPositions[j - 1] + ttCtx.dataPointsDividedWidth / 2
       if (isNaN(bcx)) {
